# stockbuddy-backend-main/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StockBuddy API",
    "description": "Comprehensive API for StockBuddy - Inventory, Transaction, and Event Management System with multi-tenant architecture, JWT authentication, and extensive business logic for managing members, vendors, banks, transactions, inventory records, events, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "StockBuddy API Support",
      "email": "support@stockbuddy.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base Path"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint. Include as: Authorization: Bearer <token>"
      }
    },
    "schemas": {
...
      "Member": {
        "type": "object",
        "description": "Member/customer record with team affiliation and membership details",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique member identifier"
          },
          "name": {
            "type": "string",
            "description": "Member's full name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Member's email address"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "BANNED"
            ],
            "description": "Member account status"
          },
          "team_name": {
            "type": "string",
            "description": "Team display name"
          },
          "membership_price": {
            "type": "number",
            "description": "Membership fee amount"
          }
        }
      },
...
      "BulkImportResponse": {
        "type": "object",
        "description": "Response for bulk import operations",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "processed": {
                "type": "integer",
                "description": "Total rows processed"
              },
              "created": {
                "type": "integer",
                "description": "Successfully created records"
              },
              "failed": {
                "type": "integer",
                "description": "Failed records"
              },
              "errors": {
                "type": "array",
                "description": "Detailed error information",
                "items": {
                  "type": "object",
                  "properties": {
                    "rowNumber": {
                      "type": "integer",
                      "description": "Row number with error"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error description"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/members": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "List Members",
        "description": "Get filtered list of organization members. Supports filtering by team, fixture, game, and search terms. Only returns ACTIVE members unless specifically filtered.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search term for member name or email",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "team_names",
            "in": "query",
            "description": "Comma-separated list of team names to filter by",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Members retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Member"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Create Member",
        "description": "Create a new member record. Team information will be resolved via API Football if team_name is provided.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "team_name",
                  "membership_price"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@example.com"
                  },
                  "team_name": {
                    "type": "string",
                    "example": "Arsenal"
                  },
                  "membership_price": {
                    "type": "number",
                    "example": 50
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Member"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/members/export": {
      "get": {
        "tags": [
          "Members"
        ],
        "summary": "Export Members to CSV",
        "description": "Export members list as CSV file with all member details. Supports template mode for downloading sample format.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "template",
            "in": "query",
            "description": "Return CSV template with sample data instead of actual members",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file with members data",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/members/import": {
      "post": {
        "tags": [
          "Members"
        ],
        "summary": "Import Members from CSV/Excel",
        "description": "Bulk import members from CSV or Excel file. Maximum 2000 rows per import. Processing is done in parallel with error reporting for failed rows.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CSV or Excel file (max 10MB). Required columns: name, email, team_name, membership_price"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import completed with results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkImportResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}